segmentation refactorings and global clustering
Now the segmentation and interpretation by googlenet
is done in one go. We can keep all data in memory, so
there is no need to write the segments to disk now.

"Checkpoint" files are created, so that we do not always
have to redo the segmentation and are able to quickly
load the dataset into memory (takes around a second
or so).