The CNN is configurable via the config
By default now the config uses resnet256.
In order for this to work, one needs to place the
model_state_{num_features}.pth files that we obtained
from Maike into the data/pre-trained-models/ folder.
Then train.py can be run. To not go through the
hassle of segmenting the images again, one can also copy
or symlink the segments folder of the non-cropped images.